requirements.txt
# Core
torch>=2.2
torchvision>=0.17
numpy>=1.26
pyyaml>=6.0

# UX & logs
rich>=13.7
tqdm>=4.66
pytz>=2024.1

# Metrics & plots (optional)
scikit-learn>=1.4
matplotlib>=3.8
